feat: 사용자 예약 삭제 시 다음 대기 예약 자동 승인 구현

+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -91,12 +91,23 @@ public class ReservationService {
 
     @Transactional(rollbackFor = Exception.class)
     public void deleteReservation(Long id) {
+        MemberReservation memberReservation = findMemberReservationById(id);
         memberReservationRepository.deleteById(id);
+
+        confirmFirstWaitingReservation(memberReservation.getReservation());
     }
 
     @Transactional(rollbackFor = Exception.class)
     public void deleteReservation(Long id, LoginMember loginMember) {
-        findMemberReservationById(id).validateIsOwner(loginMember);
+        MemberReservation memberReservation = findMemberReservationById(id);
+        memberReservation.validateIsOwner(loginMember);
         memberReservationRepository.deleteById(id);
+
+        confirmFirstWaitingReservation(memberReservation.getReservation());
+    }
+
+    private void confirmFirstWaitingReservation(Reservation reservation) {
+        memberReservationRepository.findFirstByReservationAndStatus(reservation, ReservationStatus.WAITING)
+                .ifPresent((memberReservation) -> memberReservation.setStatus(ReservationStatus.CONFIRMATION));
     }
 }
